emacs-29: Theme new mode-line-active
Emacs 29 adds a new face `mode-line-active', and uses that instead of
`mode-line' for the active mode line.  Additionally `mode-line-active'
and `mode-line-inactive' now inherit from both `mode-line' as well as
from `variable-pitch'.

Since we themed `mode-line-inactive' the inherit value from the
default value is lost and so it continued to use fixed pitch (or
rather whatever the `default' face uses.

Since we did not theme `mode-line-active', that used the default value
and thus it uses variable pitch in Emacs 29.

In this first iteration we go back to using fixed pitch for these
three faces because the switch to fixed pitch has more issues, such as
breaking my moody package.